add a joint command timeout to ModelTask
This allows to have a ModelTask behavior closer to actual hardware,
where it is common to have a timeout on the command (instead of
the current behaviour that was dependent on the Gazebo period).

The default timeout is one second. The current behaviour can be
regained by setting the timeout to zero