fix overflow problem on getCurrentTime
The way it was implemented was leading to overflow, making sim_timestamp of Line:45 be negative. It was impacting the timestamp of sensors such as GPS and IMU, making them delayed by some hours.

It caused velocity provider and other filter estimators to go to CRITICAL_ALIGNMENT_FAILURE exception state.