imu: allow changing the IMU reference pose
By default, a Gazebo IMU actually gives the transformation between
its initial pose and the current pose. This is .... not-so expected
for most of the IMUs (one would at least expect that the IMU would
correct for gravity).

In any case, provide some control over that, allowing to either retain
the default (and current) configuration, or use
REFERENCE_HORIZONTAL_PLANE to correct for gravity but keep the offset
in heading, or REFERENCE_ABSOLUTE to provide the absolute orientation.
to keep t