add support for joint scopes
The current naming scheme forces component networks to be configured
based on where a joint is attached on the whole system. Scoping the
joints to 'self' (i.e. to the model or sub-model) removes this problem,
and allow to attach a model (e.g. an arm) anywhere on a robot, while
keeping the rest of the network(s) the same.