express the accelerations in the body frame
This is really what makes the most sense from a sensing perspective.
Sensors (gyros, accelerometers) and models (hydrodynamic models)
give access to velocities and accelerations w.r.t. the inertial
frame, but expressed in the body frame.

However, expressing the velocities/accelerations in the world frame
from these sensing/estimation methods would require to have an estimate
of the system's pose in the world, which is a harder thing to get