use simulation time to determine the link state period
When _use_sim_time is false, the code would use the realtime. Because
of jitter, it would regulary do one simulation period too many if the
link export period was an exact multiple of the simulation period.

Rather use the simulation time in all cases. In case where _use_sim_time,
one should expect the simulation to run with a 1:1 time factor anyways.
This avoids implementing some complex logic to deal with the jitter.