add support to rename the task based on the gazebo plugin name
The Syskit support always assumed the task was named on the basis
of the plugin name, but the C++ integration was not actually doing
so (which would lead to collisions if there was the same plugin
spawned more than once in any given model).

This was confusing as one would basically have to make sure the plugin
name was set to whatever the task was expecting.