fix joint name validation
The current code was happily assuming that the `names` vector is of the
same size than the elements, which obviously could lead to OOB access.
Weirdly enough, on my dev machine, it would end up pointing to the
expected names (!) and pass the test. This is obviously not good.

Fix the whole validation routine, moving it to a separate method
that is called only for new elements (avoiding re-validating the
same command over and over again).

Also, add an `ignore_joint_names` flag to retain the "I don't use names"
behavior, but this time in a controlled manner.