fix bridge between Gazebo callbacks and Rock ports
This fixes two bad behaviors for some of the sensors: bad rates (lost
and/or repeated samples) and samples being sent outside of the task's
RUNNING state.

I tried to apply the same implementation pattern in all sensors, that
is avoiding to go through the updateHook, instead sending samples
straight from the callback. This is due to the (unfortunate) Gazebo
choice of using an asynchronous transport for sensors. Because of this
, the sensor samples arrive at (more or less) the right global rate,
but some jitter in logical time. In other words, one could easily receive
more than one sample in a given simulation cycle.

The mitigation used so far was to use a vector to transmit samples from
the callback to the updateHook, which is quite inefficient. Writing on
ports can be done from any thread, so do it from the callback instead.