fix: no longer expect that the joint name is a scoped name
Expecting that the user always input a scoped name increases complexity
on the robot integration side. If a robot is present in multiple
gazebo worlds, and some of these world have different scopes, one would
have to explicitly change the joint_names for each of these scopes,
e.g.:

```
robot_with_payload -> with a joint name robot_with_payload::robot::joint
standalone_robot -> with a joint name robot::joint
```

This would allow users to use a relative names when exporting a joint,
which might cause multiple exports of a joint. But then, the previous
version also allows this but with a scoped name...